Major simplification of websocket handling, reduction of threads, removal of message communication, reorganization of files in packages.
