+++ addons/binding/org.openhab.binding.loxone/META-INF/MANIFEST.MF
@@ -19,6 +19,7 @@ Import-Package:
  org.eclipse.jetty.http,
  org.eclipse.jetty.util,
  org.eclipse.jetty.util.component,
+ org.eclipse.jetty.util.thread,
  org.eclipse.jetty.websocket.api,
  org.eclipse.jetty.websocket.api.annotations,
  org.eclipse.jetty.websocket.api.extensions,
