+++ addons/binding/org.openhab.binding.loxone/src/main/java/org/openhab/binding/loxone/internal/types/LxContainer.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.loxone.internal.core;
+package org.openhab.binding.loxone.internal.types;
 
 import java.util.HashSet;
 import java.util.Set;
